Add console action for vms
