+++ code/iaas/api-logic/src/main/resources/META-INF/cattle/iaas-api/spring-iaas-api-logic-context.xml
@@ -20,7 +20,6 @@
     <bean class="io.cattle.platform.iaas.api.filter.instance.InstanceImageValidationFilter" />
     <bean class="io.cattle.platform.iaas.api.filter.instance.InstanceNetworkValidationFilter" />
     <bean class="io.cattle.platform.iaas.api.filter.instance.InstancePortsValidationFilter" />
-    <bean class="io.cattle.platform.iaas.api.instance.InstanceConsoleActionHandler" />
     <bean class="io.cattle.platform.iaas.api.credential.SshKeyOutputFilter" />
     <bean class="io.cattle.platform.iaas.api.filter.containerevent.ContainerEventFilter" />
     <bean class="io.cattle.platform.iaas.api.filter.externalevent.ExternalEventFilter" />
