+++ code/implementation/docker/api/src/main/resources/META-INF/cattle/docker-api/docker-api-defaults.properties
@@ -1,3 +1,3 @@
-console.agent.path=/v1/exec/
+exec.agent.path=/v1/exec/
 host.logs.path=/v1/logs/
 host.socketproxy.path=/v1/dockersocket/
