+++ code/implementation/vm/src/main/resources/META-INF/cattle/api-server/spring-vm-api-context.xml
@@ -6,16 +6,14 @@
     xmlns:tx="http://www.springframework.org/schema/tx"
     xmlns:extension="http://cattle.io/schemas/spring/extension"
     xsi:schemaLocation="http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop-3.0.xsd
-		http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-		http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.2.xsd
-		http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
+        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
+        http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.2.xsd
+        http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
         http://cattle.io/schemas/spring/extension http://cattle.io/schemas/spring/extension-1.0.xsd
-		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
+        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
 
     <!-- This really shouldn't be here, it's because LinkHandler and ActionHandler are
          not properly managed and rely on Spring auto wiring and not the ExtensionManager -->
-    <bean class="io.cattle.platform.docker.api.ExecActionHandler" />
-    <bean class="io.cattle.platform.docker.api.ContainerLogsActionHandler" />
-    <bean class="io.cattle.platform.docker.api.DockerSocketProxyActionHandler" />
+    <bean class="io.cattle.platform.vm.api.InstanceConsoleActionHandler" />
 
 </beans>
