+++ code/implementation/vm/src/main/resources/META-INF/cattle/api-server/vm-api-defaults.properties
@@ -1,0 +1,1 @@
+console.agent.path=/v1/console/
