+++ code/implementation/vm/src/main/resources/META-INF/cattle/core-process/spring-vm-logic-context.xml
